# Conflicts:
#	common/src/Assets/Bsp29Model.h
#	common/src/Assets/Md2Model.h
#	common/src/Assets/MdlModel.h

+++ common/src/Renderer/RenderContext.cpp
@@ -148,11 +148,11 @@ namespace TrenchBroom {
             m_showGrid = showGrid;
         }
         
-        size_t RenderContext::gridSize() const {
+        FloatType RenderContext::gridSize() const {
             return m_gridSize;
         }
         
-        void RenderContext::setGridSize(const size_t gridSize) {
+        void RenderContext::setGridSize(const FloatType gridSize) {
             m_gridSize = gridSize;
         }
